+++ include/H2DE/surfaces/H2DE_texture.h
@@ -25,6 +25,7 @@ public:
     unsigned int setColor(const H2DE_ColorRGB& color, unsigned int duration, H2DE_Easing easing, const std::function<void()>& completed, bool pauseSensitive);
 
     friend class H2DE_Object;
+    friend class H2DE_TextObject;
 };
 
 #endif